tools/test/openhst/requirements.txt
google==2.0.3
absl-py==0.9.0
protobuf==3.11.3
pexpect==4.8.0
